+++ src/aws-utilization.tsx
@@ -4,11 +4,13 @@ import { AwsService, AwsServiceUtilizationFactory } from './aws-service-utilizat
 import { getAwsCredentialsProvider } from './utils/utils.js';
 
 type AwsUtilizationType = Widget & {
+   // utilizations: Utilization[];
   awsServices: AwsService[],
   region: string
 }
 
 export class AwsUtilization extends BaseWidget {
+  // utilizations: Utilization[];
   awsServices: AwsService[];
   region: string;
 
@@ -22,11 +24,11 @@ export class AwsUtilization extends BaseWidget {
     const awsCredentialsProvider = getAwsCredentialsProvider(providers);
     for (const awsService of this.awsServices) {
       const awsServiceUtilization = AwsServiceUtilizationFactory.createObject(awsService);
-      const assessments = awsServiceUtilization.getAssessment(awsCredentialsProvider, this.region);
+      await awsServiceUtilization.getUtilization(awsCredentialsProvider, this.region, overrides);
     }
   }
   
-  render (children?: (Widget & { renderedElement: JSX.Element; })[], overridesCallback?: (overrides: any) => void): JSX.Element {
+  render (_children?: (Widget & { renderedElement: JSX.Element; })[], _overridesCallback?: (overrides: any) => void): JSX.Element {
     return <></>;
   }
 }
